Fix review notes and moved loading hint from common to core package
