+++ src/js/angular/core/loading-hint.js
