+++ src/vendor.js
@@ -6,5 +6,5 @@ import './css/lib/animate/animate.css';
 import 'jquery';
 import 'lib/angularjs/1.3.8/angular.js';
 import 'lib/bootstrap/bootstrap.min';
-import 'angular/common/loading-hint.js';
+import 'angular/core/loading-hint.js';
 import 'bootstrap-switch/dist/js/bootstrap-switch.min';
